+++ packages-deprecated/extension-placeholder/CHANGELOG.md
@@ -1,5 +1,85 @@
 # Change Log
 
+## 3.18.0
+
+### Patch Changes
+
+- Updated dependencies [a65e55d]
+  - @tiptap/extensions@3.18.0
+
+## 3.17.1
+
+### Patch Changes
+
+- @tiptap/extensions@3.17.1
+
+## 3.17.0
+
+### Patch Changes
+
+- @tiptap/extensions@3.17.0
+
+## 3.16.0
+
+### Patch Changes
+
+- @tiptap/extensions@3.16.0
+
+## 3.15.3
+
+### Patch Changes
+
+- @tiptap/extensions@3.15.3
+
+## 3.15.2
+
+### Patch Changes
+
+- @tiptap/extensions@3.15.2
+
+## 3.15.1
+
+### Patch Changes
+
+- @tiptap/extensions@3.15.1
+
+## 3.15.0
+
+### Patch Changes
+
+- @tiptap/extensions@3.15.0
+
+## 3.14.0
+
+### Patch Changes
+
+- @tiptap/extensions@3.14.0
+
+## 3.13.0
+
+### Patch Changes
+
+- @tiptap/extensions@3.13.0
+
+## 3.12.1
+
+### Patch Changes
+
+- @tiptap/extensions@3.12.1
+
+## 3.12.0
+
+### Patch Changes
+
+- @tiptap/extensions@3.12.0
+
+## 3.11.1
+
+### Patch Changes
+
+- Updated dependencies [eea7190]
+  - @tiptap/extensions@3.11.1
+
 ## 3.11.0
 
 ### Patch Changes